change size_t to long long

+++ philo/incl/philosophers.h
@@ -6,7 +6,7 @@
 /*   By: dvan-kri <dvan-kri@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2022/04/21 12:07:26 by dvan-kri      #+#    #+#                 */
-/*   Updated: 2022/05/18 10:46:42 by dvan-kri      ########   odam.nl         */
+/*   Updated: 2022/05/23 15:19:51 by dvan-kri      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -38,7 +38,7 @@ typedef struct s_philosopher
 	pthread_mutex_t	*right_fork;
 	pthread_mutex_t	last_eaten_mutex;
 	pthread_t		thread;
-	size_t			last_eaten;
+	long long		last_eaten;
 	int				id;
 	t_bool			is_dead;
 	struct s_data	*data;
@@ -53,10 +53,10 @@ typedef struct s_data
 	pthread_t		monitoring_thread;
 	t_bool			philosopher_dead;
 	t_bool			number_of_times_passed;
-	size_t			start_time;
-	size_t			time_to_die;
-	size_t			time_to_eat;
-	size_t			time_to_sleep;
+	long long		start_time;
+	long long		time_to_die;
+	long long		time_to_eat;
+	long long		time_to_sleep;
 	int				number_of_philosophers;
 	int				number_of_times_each_philosopher_must_eat;
 }	t_data;
